requirements.txt
streamlit==1.32.0
python-docx==1.1.0
chardet==5.2.0
requests==2.31.0
beautifulsoup4==4.12.3
markdown==3.5.2
langchain>=0.0.200
langchain-openai>=0.0.2
langchain-core==0.1.32
python-pptx>=0.6.21
matplotlib==3.8.3
numpy>=1.21.0
sentence-transformers==2.5.1
ipywidgets>=7.6.0
requests>=2.25.1
beautifulsoup4>=4.9.3
PyPDF2>=2.0.0
python-docx>=0.8.11
langchain-text-splitters>=0.0.1
psutil>=5.9.0
openai>=1.3.5
scikit-learn>=0.24.2
joblib>=1.0.1 